+++ scripts/rollback/rollback-3.0.13.sh
@@ -1,0 +1,276 @@
+#!/bin/bash
+# timestamp: 1646114349499
+
+npm dist-tag add @midwayjs/egg-layer@3.0.11 latest
+npm dist-tag add @midwayjs/egg-layer@3.0.13 beta
+npm dist-tag add @midwayjs/express-layer@3.0.11 latest
+npm dist-tag add @midwayjs/express-layer@3.0.13 beta
+npm dist-tag add @midwayjs/koa-layer@3.0.11 latest
+npm dist-tag add @midwayjs/koa-layer@3.0.13 beta
+npm dist-tag add @midwayjs/serverless-app@3.0.11 latest
+npm dist-tag add @midwayjs/serverless-app@3.0.13 beta
+npm dist-tag add @midwayjs/serverless-aws-starter@3.0.10 latest
+npm dist-tag add @midwayjs/serverless-aws-starter@3.0.13 beta
+npm dist-tag add @midwayjs/serverless-fc-starter@3.0.11 latest
+npm dist-tag add @midwayjs/serverless-fc-starter@3.0.13 beta
+npm dist-tag add @midwayjs/serverless-fc-trigger@3.0.11 latest
+npm dist-tag add @midwayjs/serverless-fc-trigger@3.0.13 beta
+npm dist-tag add @midwayjs/serverless-scf-trigger@3.0.11 latest
+npm dist-tag add @midwayjs/serverless-scf-trigger@3.0.13 beta
+npm dist-tag add @midwayjs/serverless-vercel-starter@3.0.11 latest
+npm dist-tag add @midwayjs/serverless-vercel-starter@3.0.13 beta
+npm dist-tag add @midwayjs/serverless-worker-starter@3.0.12 latest
+npm dist-tag add @midwayjs/serverless-worker-starter@3.0.13 beta
+npm dist-tag add @midwayjs/static-layer@3.0.11 latest
+npm dist-tag add @midwayjs/static-layer@3.0.13 beta
+npm dist-tag add @midwayjs/axios@3.0.11 latest
+npm dist-tag add @midwayjs/axios@3.0.13 beta
+npm dist-tag add @midwayjs/bootstrap@3.0.11 latest
+npm dist-tag add @midwayjs/bootstrap@3.0.13 beta
+npm dist-tag add @midwayjs/cache@3.0.11 latest
+npm dist-tag add @midwayjs/cache@3.0.13 beta
+npm dist-tag add @midwayjs/consul@3.0.11 latest
+npm dist-tag add @midwayjs/consul@3.0.13 beta
+npm dist-tag add @midwayjs/core@3.0.11 latest
+npm dist-tag add @midwayjs/core@3.0.13 beta
+npm dist-tag add @midwayjs/cos@3.0.11 latest
+npm dist-tag add @midwayjs/cos@3.0.13 beta
+npm dist-tag add @midwayjs/cross-domain@3.0.11 latest
+npm dist-tag add @midwayjs/cross-domain@3.0.13 beta
+npm dist-tag add @midwayjs/decorator@2.14.0 latest
+npm dist-tag add @midwayjs/decorator@3.0.10 beta
+npm dist-tag add @midwayjs/express-session@3.0.11 latest
+npm dist-tag add @midwayjs/express-session@3.0.13 beta
+npm dist-tag add @midwayjs/faas@3.0.11 latest
+npm dist-tag add @midwayjs/faas@3.0.13 beta
+npm dist-tag add @midwayjs/grpc@3.0.11 latest
+npm dist-tag add @midwayjs/grpc@3.0.13 beta
+npm dist-tag add @midwayjs/i18n@3.0.11 latest
+npm dist-tag add @midwayjs/i18n@3.0.13 beta
+npm dist-tag add @midwayjs/info@3.0.11 latest
+npm dist-tag add @midwayjs/info@3.0.13 beta
+npm dist-tag add @midwayjs/jwt@3.0.11 latest
+npm dist-tag add @midwayjs/jwt@3.0.13 beta
+npm dist-tag add @midwayjs/mock@3.0.11 latest
+npm dist-tag add @midwayjs/mock@3.0.13 beta
+npm dist-tag add @midwayjs/mongoose@3.0.11 latest
+npm dist-tag add @midwayjs/mongoose@3.0.13 beta
+npm dist-tag add @midwayjs/orm@3.0.11 latest
+npm dist-tag add @midwayjs/orm@3.0.13 beta
+npm dist-tag add @midwayjs/oss@3.0.11 latest
+npm dist-tag add @midwayjs/oss@3.0.13 beta
+npm dist-tag add @midwayjs/passport@3.0.11 latest
+npm dist-tag add @midwayjs/passport@3.0.13 beta
+npm dist-tag add @midwayjs/process-agent@3.0.11 latest
+npm dist-tag add @midwayjs/process-agent@3.0.13 beta
+npm dist-tag add @midwayjs/prometheus-socket-io@3.0.11 latest
+npm dist-tag add @midwayjs/prometheus-socket-io@3.0.13 beta
+npm dist-tag add @midwayjs/prometheus@3.0.11 latest
+npm dist-tag add @midwayjs/prometheus@3.0.13 beta
+npm dist-tag add @midwayjs/rabbitmq@3.0.11 latest
+npm dist-tag add @midwayjs/rabbitmq@3.0.13 beta
+npm dist-tag add @midwayjs/redis@3.0.11 latest
+npm dist-tag add @midwayjs/redis@3.0.13 beta
+npm dist-tag add @midwayjs/security@3.0.11 latest
+npm dist-tag add @midwayjs/security@3.0.13 beta
+npm dist-tag add @midwayjs/sequelize@3.0.11 latest
+npm dist-tag add @midwayjs/sequelize@3.0.13 beta
+npm dist-tag add @midwayjs/session@3.0.11 latest
+npm dist-tag add @midwayjs/session@3.0.13 beta
+npm dist-tag add @midwayjs/socketio@3.0.11 latest
+npm dist-tag add @midwayjs/socketio@3.0.13 beta
+npm dist-tag add @midwayjs/static-file@3.0.11 latest
+npm dist-tag add @midwayjs/static-file@3.0.13 beta
+npm dist-tag add @midwayjs/swagger@3.0.11 latest
+npm dist-tag add @midwayjs/swagger@3.0.13 beta
+npm dist-tag add @midwayjs/tablestore@3.0.11 latest
+npm dist-tag add @midwayjs/tablestore@3.0.13 beta
+npm dist-tag add @midwayjs/task@3.0.11 latest
+npm dist-tag add @midwayjs/task@3.0.13 beta
+npm dist-tag add @midwayjs/typegoose@3.0.11 latest
+npm dist-tag add @midwayjs/typegoose@3.0.13 beta
+npm dist-tag add @midwayjs/upload@3.0.11 latest
+npm dist-tag add @midwayjs/upload@3.0.13 beta
+npm dist-tag add @midwayjs/validate@3.0.11 latest
+npm dist-tag add @midwayjs/validate@3.0.13 beta
+npm dist-tag add @midwayjs/version@3.0.12 latest
+npm dist-tag add @midwayjs/version@3.0.13 beta
+npm dist-tag add @midwayjs/view-ejs@3.0.11 latest
+npm dist-tag add @midwayjs/view-ejs@3.0.13 beta
+npm dist-tag add @midwayjs/view-nunjucks@3.0.11 latest
+npm dist-tag add @midwayjs/view-nunjucks@3.0.13 beta
+npm dist-tag add @midwayjs/view@3.0.11 latest
+npm dist-tag add @midwayjs/view@3.0.13 beta
+npm dist-tag add @midwayjs/express@3.0.11 latest
+npm dist-tag add @midwayjs/express@3.0.13 beta
+npm dist-tag add @midwayjs/koa@3.0.11 latest
+npm dist-tag add @midwayjs/koa@3.0.13 beta
+npm dist-tag add @midwayjs/web@3.0.11 latest
+npm dist-tag add @midwayjs/web@3.0.13 beta
+npm dist-tag add @midwayjs/ws@3.0.11 latest
+npm dist-tag add @midwayjs/ws@3.0.13 beta
+tnpm dist-tag add @midwayjs/egg-layer@3.0.11 latest
+tnpm dist-tag add @midwayjs/egg-layer@3.0.13 beta
+tnpm dist-tag add @midwayjs/express-layer@3.0.11 latest
+tnpm dist-tag add @midwayjs/express-layer@3.0.13 beta
+tnpm dist-tag add @midwayjs/koa-layer@3.0.11 latest
+tnpm dist-tag add @midwayjs/koa-layer@3.0.13 beta
+tnpm dist-tag add @midwayjs/serverless-app@3.0.11 latest
+tnpm dist-tag add @midwayjs/serverless-app@3.0.13 beta
+tnpm dist-tag add @midwayjs/serverless-aws-starter@3.0.10 latest
+tnpm dist-tag add @midwayjs/serverless-aws-starter@3.0.13 beta
+tnpm dist-tag add @midwayjs/serverless-fc-starter@3.0.11 latest
+tnpm dist-tag add @midwayjs/serverless-fc-starter@3.0.13 beta
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@3.0.11 latest
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@3.0.13 beta
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@3.0.11 latest
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@3.0.13 beta
+tnpm dist-tag add @midwayjs/serverless-vercel-starter@3.0.11 latest
+tnpm dist-tag add @midwayjs/serverless-vercel-starter@3.0.13 beta
+tnpm dist-tag add @midwayjs/serverless-worker-starter@3.0.12 latest
+tnpm dist-tag add @midwayjs/serverless-worker-starter@3.0.13 beta
+tnpm dist-tag add @midwayjs/static-layer@3.0.11 latest
+tnpm dist-tag add @midwayjs/static-layer@3.0.13 beta
+tnpm dist-tag add @midwayjs/axios@3.0.11 latest
+tnpm dist-tag add @midwayjs/axios@3.0.13 beta
+tnpm dist-tag add @midwayjs/bootstrap@3.0.11 latest
+tnpm dist-tag add @midwayjs/bootstrap@3.0.13 beta
+tnpm dist-tag add @midwayjs/cache@3.0.11 latest
+tnpm dist-tag add @midwayjs/cache@3.0.13 beta
+tnpm dist-tag add @midwayjs/consul@3.0.11 latest
+tnpm dist-tag add @midwayjs/consul@3.0.13 beta
+tnpm dist-tag add @midwayjs/core@3.0.11 latest
+tnpm dist-tag add @midwayjs/core@3.0.13 beta
+tnpm dist-tag add @midwayjs/cos@3.0.11 latest
+tnpm dist-tag add @midwayjs/cos@3.0.13 beta
+tnpm dist-tag add @midwayjs/cross-domain@3.0.11 latest
+tnpm dist-tag add @midwayjs/cross-domain@3.0.13 beta
+tnpm dist-tag add @midwayjs/decorator@2.14.0 latest
+tnpm dist-tag add @midwayjs/decorator@3.0.10 beta
+tnpm dist-tag add @midwayjs/express-session@3.0.11 latest
+tnpm dist-tag add @midwayjs/express-session@3.0.13 beta
+tnpm dist-tag add @midwayjs/faas@3.0.11 latest
+tnpm dist-tag add @midwayjs/faas@3.0.13 beta
+tnpm dist-tag add @midwayjs/grpc@3.0.11 latest
+tnpm dist-tag add @midwayjs/grpc@3.0.13 beta
+tnpm dist-tag add @midwayjs/i18n@3.0.11 latest
+tnpm dist-tag add @midwayjs/i18n@3.0.13 beta
+tnpm dist-tag add @midwayjs/info@3.0.11 latest
+tnpm dist-tag add @midwayjs/info@3.0.13 beta
+tnpm dist-tag add @midwayjs/jwt@3.0.11 latest
+tnpm dist-tag add @midwayjs/jwt@3.0.13 beta
+tnpm dist-tag add @midwayjs/mock@3.0.11 latest
+tnpm dist-tag add @midwayjs/mock@3.0.13 beta
+tnpm dist-tag add @midwayjs/mongoose@3.0.11 latest
+tnpm dist-tag add @midwayjs/mongoose@3.0.13 beta
+tnpm dist-tag add @midwayjs/orm@3.0.11 latest
+tnpm dist-tag add @midwayjs/orm@3.0.13 beta
+tnpm dist-tag add @midwayjs/oss@3.0.11 latest
+tnpm dist-tag add @midwayjs/oss@3.0.13 beta
+tnpm dist-tag add @midwayjs/passport@3.0.11 latest
+tnpm dist-tag add @midwayjs/passport@3.0.13 beta
+tnpm dist-tag add @midwayjs/process-agent@3.0.11 latest
+tnpm dist-tag add @midwayjs/process-agent@3.0.13 beta
+tnpm dist-tag add @midwayjs/prometheus-socket-io@3.0.11 latest
+tnpm dist-tag add @midwayjs/prometheus-socket-io@3.0.13 beta
+tnpm dist-tag add @midwayjs/prometheus@3.0.11 latest
+tnpm dist-tag add @midwayjs/prometheus@3.0.13 beta
+tnpm dist-tag add @midwayjs/rabbitmq@3.0.11 latest
+tnpm dist-tag add @midwayjs/rabbitmq@3.0.13 beta
+tnpm dist-tag add @midwayjs/redis@3.0.11 latest
+tnpm dist-tag add @midwayjs/redis@3.0.13 beta
+tnpm dist-tag add @midwayjs/security@3.0.11 latest
+tnpm dist-tag add @midwayjs/security@3.0.13 beta
+tnpm dist-tag add @midwayjs/sequelize@3.0.11 latest
+tnpm dist-tag add @midwayjs/sequelize@3.0.13 beta
+tnpm dist-tag add @midwayjs/session@3.0.11 latest
+tnpm dist-tag add @midwayjs/session@3.0.13 beta
+tnpm dist-tag add @midwayjs/socketio@3.0.11 latest
+tnpm dist-tag add @midwayjs/socketio@3.0.13 beta
+tnpm dist-tag add @midwayjs/static-file@3.0.11 latest
+tnpm dist-tag add @midwayjs/static-file@3.0.13 beta
+tnpm dist-tag add @midwayjs/swagger@3.0.11 latest
+tnpm dist-tag add @midwayjs/swagger@3.0.13 beta
+tnpm dist-tag add @midwayjs/tablestore@3.0.11 latest
+tnpm dist-tag add @midwayjs/tablestore@3.0.13 beta
+tnpm dist-tag add @midwayjs/task@3.0.11 latest
+tnpm dist-tag add @midwayjs/task@3.0.13 beta
+tnpm dist-tag add @midwayjs/typegoose@3.0.11 latest
+tnpm dist-tag add @midwayjs/typegoose@3.0.13 beta
+tnpm dist-tag add @midwayjs/upload@3.0.11 latest
+tnpm dist-tag add @midwayjs/upload@3.0.13 beta
+tnpm dist-tag add @midwayjs/validate@3.0.11 latest
+tnpm dist-tag add @midwayjs/validate@3.0.13 beta
+tnpm dist-tag add @midwayjs/version@3.0.12 latest
+tnpm dist-tag add @midwayjs/version@3.0.13 beta
+tnpm dist-tag add @midwayjs/view-ejs@3.0.11 latest
+tnpm dist-tag add @midwayjs/view-ejs@3.0.13 beta
+tnpm dist-tag add @midwayjs/view-nunjucks@3.0.11 latest
+tnpm dist-tag add @midwayjs/view-nunjucks@3.0.13 beta
+tnpm dist-tag add @midwayjs/view@3.0.11 latest
+tnpm dist-tag add @midwayjs/view@3.0.13 beta
+tnpm dist-tag add @midwayjs/express@3.0.11 latest
+tnpm dist-tag add @midwayjs/express@3.0.13 beta
+tnpm dist-tag add @midwayjs/koa@3.0.11 latest
+tnpm dist-tag add @midwayjs/koa@3.0.13 beta
+tnpm dist-tag add @midwayjs/web@3.0.11 latest
+tnpm dist-tag add @midwayjs/web@3.0.13 beta
+tnpm dist-tag add @midwayjs/ws@3.0.11 latest
+tnpm dist-tag add @midwayjs/ws@3.0.13 beta
+
+# Changes:
+
+#  - @midwayjs/egg-layer: 3.0.11 => 3.0.13
+#  - @midwayjs/express-layer: 3.0.11 => 3.0.13
+#  - @midwayjs/koa-layer: 3.0.11 => 3.0.13
+#  - @midwayjs/serverless-app: 3.0.11 => 3.0.13
+#  - @midwayjs/serverless-aws-starter: 3.0.10 => 3.0.13
+#  - @midwayjs/serverless-fc-starter: 3.0.11 => 3.0.13
+#  - @midwayjs/serverless-fc-trigger: 3.0.11 => 3.0.13
+#  - @midwayjs/serverless-scf-trigger: 3.0.11 => 3.0.13
+#  - @midwayjs/serverless-vercel-starter: 3.0.11 => 3.0.13
+#  - @midwayjs/serverless-worker-starter: 3.0.12 => 3.0.13
+#  - @midwayjs/static-layer: 3.0.11 => 3.0.13
+#  - @midwayjs/axios: 3.0.11 => 3.0.13
+#  - @midwayjs/bootstrap: 3.0.11 => 3.0.13
+#  - @midwayjs/cache: 3.0.11 => 3.0.13
+#  - @midwayjs/consul: 3.0.11 => 3.0.13
+#  - @midwayjs/core: 3.0.11 => 3.0.13
+#  - @midwayjs/cos: 3.0.11 => 3.0.13
+#  - @midwayjs/cross-domain: 3.0.11 => 3.0.13
+#  - @midwayjs/decorator: 2.14.0 => 3.0.13
+#  - @midwayjs/express-session: 3.0.11 => 3.0.13
+#  - @midwayjs/faas: 3.0.11 => 3.0.13
+#  - @midwayjs/grpc: 3.0.11 => 3.0.13
+#  - @midwayjs/i18n: 3.0.11 => 3.0.13
+#  - @midwayjs/info: 3.0.11 => 3.0.13
+#  - @midwayjs/jwt: 3.0.11 => 3.0.13
+#  - @midwayjs/mock: 3.0.11 => 3.0.13
+#  - @midwayjs/mongoose: 3.0.11 => 3.0.13
+#  - @midwayjs/orm: 3.0.11 => 3.0.13
+#  - @midwayjs/oss: 3.0.11 => 3.0.13
+#  - @midwayjs/passport: 3.0.11 => 3.0.13
+#  - @midwayjs/process-agent: 3.0.11 => 3.0.13
+#  - @midwayjs/prometheus-socket-io: 3.0.11 => 3.0.13
+#  - @midwayjs/prometheus: 3.0.11 => 3.0.13
+#  - @midwayjs/rabbitmq: 3.0.11 => 3.0.13
+#  - @midwayjs/redis: 3.0.11 => 3.0.13
+#  - @midwayjs/security: 3.0.11 => 3.0.13
+#  - @midwayjs/sequelize: 3.0.11 => 3.0.13
+#  - @midwayjs/session: 3.0.11 => 3.0.13
+#  - @midwayjs/socketio: 3.0.11 => 3.0.13
+#  - @midwayjs/static-file: 3.0.11 => 3.0.13
+#  - @midwayjs/swagger: 3.0.11 => 3.0.13
+#  - @midwayjs/tablestore: 3.0.11 => 3.0.13
+#  - @midwayjs/task: 3.0.11 => 3.0.13
+#  - @midwayjs/typegoose: 3.0.11 => 3.0.13
+#  - @midwayjs/upload: 3.0.11 => 3.0.13
+#  - @midwayjs/validate: 3.0.11 => 3.0.13
+#  - @midwayjs/version: 3.0.12 => 3.0.13
+#  - @midwayjs/view-ejs: 3.0.11 => 3.0.13
+#  - @midwayjs/view-nunjucks: 3.0.11 => 3.0.13
+#  - @midwayjs/view: 3.0.11 => 3.0.13
+#  - @midwayjs/express: 3.0.11 => 3.0.13
+#  - @midwayjs/koa: 3.0.11 => 3.0.13
+#  - @midwayjs/web: 3.0.11 => 3.0.13
+#  - @midwayjs/ws: 3.0.11 => 3.0.13